VRAM overflow easy fix
This is not a "proper" fix that guarantees the problem can never occur. Instead, because the system is already robust (see notes at head of idleSprites\main.event); we're fixing it by taking the significant outliers and just shrinking them.

The Fighter sprite is made by Alusq; I made all the rest of the edits.